backend/pyproject.toml
[project]
name = "invoice-trading"
version = "0.1.0"
description = ""
authors = [
    {name = "NDPhuu",email = "ducphu.tc10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "fastapi-mail (>=1.6.1,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "supabase (>=2.27.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
